fix mcp-server
