--- packages/githru-mcp-server/.gitignore
@@ -1,3 +1,0 @@
-.smithery
-node_modules
-.env
