--- packages/githru-mcp-server/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "smithery-scaffold",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"module": "./src/index.ts",
-	"type": "module",
-	"scripts": {
-		"dev": "npx @smithery/cli dev"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"dependencies": {
-		"@modelcontextprotocol/sdk": "^1.12.1",
-		"zod": "^3.25.46"
-	},
-	"devDependencies": {
-		"@smithery/cli": "^1.2.4",
-		"tsx": "^4.19.4"
-	}
-}
