--- packages/githru-mcp-server/smithery.yaml
@@ -1,1 +1,0 @@
-runtime: typescript
