+++ packages/githru-mcp/.gitignore
@@ -1,3 +1,7 @@
+.smithery
+node_modules
+.env
+
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 /dist
