+++ packages/githru-mcp/package.json
@@ -1,18 +1,22 @@
 {
-    "name": "githru-mcp",
-    "version": "0.0.1",
-    "type": "module",
-    "module": "src/index.ts",
-    "scripts": {
-        "build": "npx @smithery/cli build",
-        "dev": "npx @smithery/cli dev"
-    },
-    "dependencies": {
-        "@modelcontextprotocol/sdk": "^1.17.0",
-        "zod": "^3.25.46"
-    },
-    "devDependencies": {
-        "typescript": "^5.5.4",
-        "ts-node": "^10.9.2"
-    }
+	"name": "smithery-scaffold",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"module": "./src/index.ts",
+	"type": "module",
+	"scripts": {
+		"dev": "npx @smithery/cli dev"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"@modelcontextprotocol/sdk": "^1.12.1",
+		"zod": "^3.25.46"
+	},
+	"devDependencies": {
+		"@smithery/cli": "^1.2.4",
+		"tsx": "^4.19.4"
+	}
 }
